adding step to create artifacts folder in case it doesn't exists and copy html template to it

+++ fetch_pagespeed_json.rb
@@ -1,4 +1,5 @@
 require 'open-uri'
+require 'FileUtils'
 api_url = 'https://www.googleapis.com/pagespeedonline/v1/runPagespeed'
 api_key = 'AIzaSyCYnhJjkzR4zz_PlXbjTVoZyUESBoe6Ajc'
 callback = 'runPagespeedCallbacks'
@@ -10,7 +11,13 @@ url = [
     'callback=' + callback,
     'key=' + api_key,
   ].join('&');
-
-open('artifacts/pagespeed_results.js', 'wb') do |file|
-  file << open(url).read
-end
+  
+Dir.mkdir('artifacts') if File::directory?('artifacts') == false
+if File::exists?('pagespeed_results.html')
+	FileUtils.cp('pagespeed_results.html','artifacts/pagespeed_results.html')
+	open('artifacts/pagespeed_results.js', 'wb') do |file|
+  		file << open(url).read
+  	end	
+else
+	puts "Missing HTML Template!"
+end
